# Conflicts:
#	src/main/fe-medicareNetwork/src/board/SupplementList.js

+++ src/main/java/be/com/bemedicare/board/repository/Board1Repository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Board1Repository extends JpaRepository<Board1, Integer> {
+public interface Board1Repository extends JpaRepository<Board1, Long> {
 
 Page<Board1> findByTitleContaining(String searchKeyword, Pageable pageable);
