+++ src/main/java/be/com/bemedicare/board/repository/Board3Repository.java
@@ -5,6 +5,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Board3Repository extends JpaRepository<Board3, Integer> {
+public interface Board3Repository extends JpaRepository<Board3, Long> {
     Page<Board3> findByTitleContaining(String searchKeyword, Pageable pageable);
 }
